Rename qt_application, fix start of trace
