+++ napari/_qt/qt_timing.py
